fix validation & escapes

+++ spring/src/main/java/itmo/ivank/soa/dto/filter/CoordinatesFilter.java
@@ -4,7 +4,7 @@ import itmo.ivank.soa.dto.filter.primitive.NumberFilter;
 import jakarta.validation.Valid;
 
 public record CoordinatesFilter(
-        @Valid NumberFilter<Integer> xFilter,
-        @Valid NumberFilter<Long> yFilter
+        @Valid NumberFilter<Long> xFilter,
+        @Valid NumberFilter<Float> yFilter
 ) {
 }
